% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeNet.R
\name{CeNetOmnibus}
\alias{CeNetOmnibus}
\title{Run CeNetOmnibus App}
\usage{
CeNetOmnibus(
  maxRequestSize = 1000,
  workpath = tempdir(),
  projectName = NULL,
  typeLimit = 10,
  ...
)
}
\arguments{
\item{maxRequestSize}{Integer.The admitted file size for uploaded files. Unit MB. Default 1000MB.}

\item{workpath}{Character. The dictionary for temp files. Default is return value of tempdir()}

\item{projectName}{Character. The name of this analysis. Default is session token.}

\item{typeLimit}{Integer. The number of valid items. Default is 10.}

\item{...}{Other parameters passed to runApp()}
}
\value{

}
\description{
Run CeNetOmnibus App
}
\examples{
   CeNetOmnibus()
}
