% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependency.R
\name{checkDependency}
\alias{checkDependency}
\title{Check and Install the dependency packages of CeNetOmnibus}
\usage{
checkDependency()
}
\value{

}
\description{
Check and Install the dependency packages of CeNetOmnibus
}
\examples{
   checkDependency()
}
